Frankreed/unify result screens (#96)
* Result screens unified!

* fixed the display bug in result

* added title to styles, some redundancy removed from barChartScreen

---------
